レポートに入出力結果を書くことにした．k03を少し補足
