+++ k00/README.tex.md
@@ -32,5 +32,14 @@ l.39: 最大値max_valを出力
 
 l.40: 最小値min_valを出力
 
+## 出力結果
+
+```
+input the filename of sample:../sample/heights_male.csv
+the filename of sample: ../sample/heights_male.csv
+Highest: 183.870000
+Lowest: 167.830000
+```
+
 ## 修正履歴
