+++ k03/README.tex.md
@@ -27,7 +27,66 @@ aa83988848 薗田光太郎
   } 
   ```
 
+### コマンドライン入力
+
+また，今回のプログラムは，実行してから必要な値を入力するのではなく，
+
+```
+./k03 170.8 5.43 5
+```
+
+などと，実行時に引数でそれらを指定するようにする．そのために，main関数の引数をvoidではなく，
+
+```
+int main(int argc, char* argv[])
+```
+
+などとする．こうすると，argcに実行時のコマンド行にあるスペースで区切られた文字列の数（今の例だと`4`）が入り，
+`argv[0]`には，`"./k03"`，`argv[1]`には，`"170.8"`，...が格納される．`sscanf`でこれらの文字列から数値を読み取ればよい．
+
+### マクロ定義による分岐
+
+コード中のコメントアウトするとき，`/*`と`*/`で囲む，もしくは `//`で行コメントアウトが使えるが，
+他の方法も知っておくとよい．
+
+1.`if(0)`で囲んだ分岐ブロックは，コメントアウトと同様に常にスキップされる．
+2.`#if 0`~`#endif`で囲んだブロックは，プリプロセスでスキップされる．
+
+違いは，前者の場合にはコード解析されるが，後者はコード解析されない．前者の場合はコメントアウトされていても，ブロックの中の文法が正しいことが要求される．
+
+いずれにせよ，if分岐がコメントアウトに使えるということだが，逆に上述の2のようなマクロは分岐に使うことができるわけで，コンパイル時に"-DTEST"を付けたときだけprintfさせたいときは，
+
+```
+#ifdef TEST
+ printf("%d\n",x);
+#endif
+```
+
+のような行を加えることができる．
+
+
 ## ソースコードの説明
+
+## 入出力結果
+
+以下は例です． コンパイルするときに`-DCONST_SEED`をつけるとこうなります．
+
+```
+============================================
+template mean: 170.8
+template standard deviation: 5.43
+Num of dummy data: 5
+============================================
+162.20
+171.29
+175.88
+168.59
+166.74
+```
+
+コンパイル時に`-DCONST_SEED`をつけない場合を以下に示す．
+
+
 
 ## 修正履歴
