+++ k03/k03.c
@@ -2,56 +2,50 @@
 #include <stdlib.h>
 #include <string.h>
 #include <math.h>
+#ifdef CONST_SEED
+#define RAND_SEED 2020
+#else
+#include <time.h>
+#define RAND_SEED ((unsigned) time(NULL))
+#endif
 
 extern double r_unif(void);
-extern double r_norm(void);
+extern double r_stdnorm(void);
 
-int main(void)
+int main(int argc, char* argv[])
 {
-    double val;
-    char fname[FILENAME_MAX];
-    char buf[256];
-    FILE* fp;
+    int num_dummy;
+    double mu;
+    double sigma;
+    int i;
+    double dummy;
 
-    printf("input the filename of sample:");
-    fgets(fname,sizeof(fname),stdin);
-    fname[strlen(fname)-1] = '\0';
-    printf("the filename of sample: %s\n",fname);
 
-    fp = fopen(fname,"r");
-    if(fp==NULL){
-        fputs("File open error\n",stderr);
-        exit(EXIT_FAILURE);
+
+
+
+    printf("============================================\n");
+    printf("template mean: %4.1lf\n",mu);
+    printf("template standard deviation: %3.2lf\n",sigma);
+    printf("Num of dummy data: %d\n",num_dummy);
+    printf("============================================\n");
+
+    srand(RAND_SEED);
+    for(  ){
+        /* r_stdnormを使って，1人のデータを捏造 */
+        dummy =   ;
+        printf("%5.2lf\n",dummy);
     }
 
-    while(fgets(buf,sizeof(buf),fp) != NULL){
-        sscanf(buf,"%lf",&val);
-
-
-    
-
-
-
-    }
-
-    if(fclose(fp) == EOF){
-        fputs("file close error\n",stderr);
-        exit(EXIT_FAILURE);
-    }
-
-
-    return 0;
-
-
+    return EXIT_SUCCESS;
 }
 
 double r_unif(void)
 {
     return (double)(rand()+1)/(RAND_MAX+2);
 }
-  
-double r_norm(void)
-{  
-    return sqrt( -2.0*log(r_unif()) ) * sin( 2.0*M_PI*r_unif() );
-} 
 
+double r_stdnorm(void)
+{
+    return sqrt( -2.0*log(r_unif()) ) * sin( 2.0*M_PI*r_unif() );
+}
