+++ k04/README.tex.md
@@ -13,5 +13,8 @@ aa83988848 薗田光太郎
 
 ## ソースコードの説明
 
+## 入出力結果
+
+
 ## 修正履歴
