+++ k04/k04.c
@@ -1,0 +1,43 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <math.h>
+
+int main(void)
+{
+    double val;
+    char fname[FILENAME_MAX];
+    char buf[256];
+    FILE* fp;
+
+    printf("input the filename of sample:");
+    fgets(fname,sizeof(fname),stdin);
+    fname[strlen(fname)-1] = '\0';
+    printf("the filename of sample: %s\n",fname);
+
+    fp = fopen(fname,"r");
+    if(fp==NULL){
+        fputs("File open error\n",stderr);
+        exit(EXIT_FAILURE);
+    }
+
+    while(fgets(buf,sizeof(buf),fp) != NULL){
+        sscanf(buf,"%lf",&val);
+
+
+    
+
+
+
+    }
+
+    if(fclose(fp) == EOF){
+        fputs("file close error\n",stderr);
+        exit(EXIT_FAILURE);
+    }
+
+
+    return 0;
+
+
+}
