+++ k05/README.tex.md
@@ -10,5 +10,7 @@ aa83988848 薗田光太郎
 
 ## ソースコードの説明
 
+## 入出力結果
+
 ## 修正履歴
